The singleton pattern is no longer used in this project.
