+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>dansplugins</groupId>
     <artifactId>Democracy</artifactId>
-    <version>0.1-alpha-1</version>
+    <version>0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>Democracy</name>
@@ -77,7 +77,7 @@
         <dependency>
             <groupId>com.github.Preponderous-Software</groupId>
             <artifactId>ponder</artifactId>
-            <version>1.0</version>
+            <version>1.1</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
