+++ src/main/java/dansplugins/democracy/commands/DefaultCommand.java
@@ -13,14 +13,16 @@ import java.util.Arrays;
  * @author Daniel McCoy Stephenson
  */
 public class DefaultCommand extends AbstractPluginCommand {
+    private final Democracy democracy;
 
-    public DefaultCommand() {
+    public DefaultCommand(Democracy democracy) {
         super(new ArrayList<>(Arrays.asList("default")), new ArrayList<>(Arrays.asList("d.default")));
+        this.democracy = democracy;
     }
 
     @Override
     public boolean execute(CommandSender commandSender) {
-        commandSender.sendMessage(ChatColor.AQUA + "Democracy " + Democracy.getInstance().getVersion());
+        commandSender.sendMessage(ChatColor.AQUA + "Democracy " + democracy.getVersion());
         commandSender.sendMessage(ChatColor.AQUA + "Developed by: Daniel Stephenson");
         commandSender.sendMessage(ChatColor.AQUA + "Wiki: https://github.com/Dans-Plugins/Democracy/wiki");
         return true;
