+++ src/main/java/dansplugins/democracy/commands/StartCommand.java
@@ -1,11 +1,11 @@
 package dansplugins.democracy.commands;
 
+import dansplugins.democracy.factories.ElectionFactory;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
 import dansplugins.democracy.Democracy;
-import dansplugins.democracy.factories.ElectionFactory;
 import dansplugins.factionsystem.externalapi.MF_Faction;
 import preponderous.ponder.minecraft.bukkit.abs.AbstractPluginCommand;
 
@@ -18,9 +18,13 @@ import java.util.UUID;
  * @author Daniel McCoy Stephenson
  */
 public class StartCommand extends AbstractPluginCommand {
+    private final Democracy democracy;
+    private final ElectionFactory electionFactory;
 
-    public StartCommand() {
+    public StartCommand(Democracy democracy, ElectionFactory electionFactory) {
         super(new ArrayList<>(Arrays.asList("start")), new ArrayList<>(Arrays.asList("d.start")));
+        this.democracy = democracy;
+        this.electionFactory = electionFactory;
     }
 
     @Override
@@ -31,13 +35,13 @@ public class StartCommand extends AbstractPluginCommand {
         }
         Player player = (Player) commandSender;
 
-        MF_Faction faction = Democracy.getInstance().getMedievalFactionsAPI().getFaction(player);
+        MF_Faction faction = democracy.getMedievalFactionsAPI().getFaction(player);
         if (faction == null || !faction.getOwner().equals(player.getUniqueId())) {
             player.sendMessage(ChatColor.RED + "You must be the owner of a faction to start an election.");
             return false;
         }
 
-        UUID electionUUID = ElectionFactory.getInstance().createElection(player);
+        UUID electionUUID = electionFactory.createElection(player);
         if (electionUUID == null) {
             player.sendMessage(ChatColor.RED + "An election is already in progress.");
             return false;
