--- src/main/java/dansplugins/democracy/eventhandlers/JoinHandler.java
@@ -1,18 +1,0 @@
-package dansplugins.democracy.eventhandlers;
-
-import org.bukkit.ChatColor;
-import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerJoinEvent;
-
-/**
- * @author Daniel McCoy Stephenson
- */
-public class JoinHandler implements Listener {
-    @EventHandler()
-    public void handle(PlayerJoinEvent event) {
-        Player player = event.getPlayer();
-        player.sendMessage(ChatColor.AQUA + "This message was sent by ExamplePonderPlugin.");
-    }
-}
