+++ src/main/java/dansplugins/democracy/listeners/JoinListener.java
@@ -1,0 +1,19 @@
+package dansplugins.democracy.listeners;
+
+import org.bukkit.ChatColor;
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerJoinEvent;
+
+/**
+ * @author Daniel McCoy Stephenson
+ */
+public class JoinListener implements Listener {
+
+    @EventHandler()
+    public void handle(PlayerJoinEvent event) {
+        Player player = event.getPlayer();
+        player.sendMessage(ChatColor.AQUA + "This message was sent by ExamplePonderPlugin.");
+    }
+}
