+++ src/main/java/dansplugins/democracy/services/ConfigService.java
@@ -1,0 +1,126 @@
+package dansplugins.democracy.services;
+
+/*
+    To add a new config option, the following methods must be altered:
+    - saveMissingConfigDefaultsIfNotPresent
+    - setConfigOption()
+    - sendConfigList()
+ */
+
+import dansplugins.democracy.Democracy;
+import org.bukkit.ChatColor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.configuration.file.FileConfiguration;
+
+/**
+ * @author Daniel McCoy Stephenson
+ */
+public class ConfigService {
+    private final Democracy democracy;
+
+    private boolean altered = false;
+
+    public ConfigService(Democracy democracy) {
+        this.democracy = democracy;
+    }
+
+    public void saveMissingConfigDefaultsIfNotPresent() {
+        // set version
+        if (!getConfig().isString("version")) {
+            getConfig().addDefault("version", democracy.getVersion());
+        } else {
+            getConfig().set("version", democracy.getVersion());
+        }
+
+        // save config options
+        if (!isSet("debugMode")) { getConfig().set("debugMode", false); }
+
+        getConfig().options().copyDefaults(true);
+        democracy.saveConfig();
+    }
+
+    public void setConfigOption(String option, String value, CommandSender sender) {
+        if (getConfig().isSet(option)) {
+            if (option.equalsIgnoreCase("version")) {
+                sender.sendMessage(ChatColor.RED + "Cannot set version.");
+                return;
+            } else if (option.equalsIgnoreCase("A")) {
+                getConfig().set(option, Integer.parseInt(value));
+                sender.sendMessage(ChatColor.GREEN + "Integer set.");
+            } else if (option.equalsIgnoreCase("debugMode")) {
+                getConfig().set(option, Boolean.parseBoolean(value));
+                sender.sendMessage(ChatColor.GREEN + "Boolean set.");
+            } else if (option.equalsIgnoreCase("C")) {
+                getConfig().set(option, Double.parseDouble(value));
+                sender.sendMessage(ChatColor.GREEN + "Double set.");
+            } else {
+                getConfig().set(option, value);
+                sender.sendMessage(ChatColor.GREEN + "String set.");
+            }
+
+            // save
+            democracy.saveConfig();
+            altered = true;
+        } else {
+            sender.sendMessage(ChatColor.RED + "That config option wasn't found.");
+        }
+    }
+
+    public void sendConfigList(CommandSender sender) {
+        sender.sendMessage(ChatColor.AQUA + "=== Config List ===");
+        sender.sendMessage(ChatColor.AQUA + "version: " + getConfig().getString("version")
+                + ", debugMode: " + getString("debugMode"));
+    }
+
+    public boolean hasBeenAltered() {
+        return altered;
+    }
+
+    public FileConfiguration getConfig() {
+        return democracy.getConfig();
+    }
+
+    public boolean isSet(String option) {
+        return getConfig().isSet(option);
+    }
+
+    public int getInt(String option) {
+        return getConfig().getInt(option);
+    }
+
+    public int getIntOrDefault(String option, int defaultValue) {
+        int toReturn = getInt(option);
+        if (toReturn == 0) {
+            return defaultValue;
+        }
+        return toReturn;
+    }
+
+    public boolean getBoolean(String option) {
+        return getConfig().getBoolean(option);
+    }
+
+    public double getDouble(String option) {
+        return getConfig().getDouble(option);
+    }
+
+    public double getDoubleOrDefault(String option, double defaultValue) {
+        double toReturn = getDouble(option);
+        if (toReturn == 0) {
+            return defaultValue;
+        }
+        return toReturn;
+    }
+
+    public String getString(String option) {
+        return getConfig().getString(option);
+    }
+
+    public String getStringOrDefault(String option, String defaultValue) {
+        String toReturn = getString(option);
+        if (toReturn == null) {
+            return defaultValue;
+        }
+        return toReturn;
+    }
+}
