+++ src/main/java/dansplugins/democracy/utils/Logger.java
@@ -6,22 +6,14 @@ import dansplugins.democracy.Democracy;
  * @author Daniel McCoy Stephenson
  */
 public class Logger {
+    private final Democracy democracy;
 
-    private static Logger instance;
-
-    private Logger() {
-
-    }
-
-    public static Logger getInstance() {
-        if (instance == null) {
-            instance = new Logger();
-        }
-        return instance;
+    public Logger(Democracy democracy) {
+        this.democracy = democracy;
     }
 
     public void log(String message) {
-        if (Democracy.getInstance().isDebugEnabled()) {
+        if (democracy.isDebugEnabled()) {
             System.out.println("[ExamplePonderPlugin] " + message);
         }
     }
